refactor: move maestro readonly bundles resource specific tests to package level ones
