+++ backend/pkg/controllers/read_and_persist_nodepool_scoped_maestro_readonly_bundles_content_controller_test.go
@@ -18,24 +18,15 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"net/http"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/mock/gomock"
-	workv1 "open-cluster-management.io/api/work/v1"
 
-	k8serrors "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 
 	azcorearm "github.com/Azure/azure-sdk-for-go/sdk/azcore/arm"
-	"github.com/Azure/azure-sdk-for-go/sdk/data/azcosmos"
-
-	hsv1beta1 "github.com/openshift/hypershift/api/hypershift/v1beta1"
 
 	"github.com/Azure/ARO-HCP/backend/pkg/controllers/controllerutils"
 	"github.com/Azure/ARO-HCP/backend/pkg/maestro"
@@ -72,311 +63,6 @@ func (e *errorInjectingSPNPCRUD) Get(ctx context.Context, resourceID string) (*a
 		return nil, e.getErr
 	}
 	return e.ServiceProviderNodePoolCRUD.Get(ctx, resourceID)
-}
-
-// errorInjectingMCCCRUDForNodePool wraps ManagementClusterContentCRUD to allow error injection.
-type errorInjectingMCCCRUDForNodePool struct {
-	database.ManagementClusterContentCRUD
-	getResult  *api.ManagementClusterContent
-	getErr     error
-	replaceErr error
-}
-
-func (e *errorInjectingMCCCRUDForNodePool) Get(ctx context.Context, resourceID string) (*api.ManagementClusterContent, error) {
-	if e.getErr != nil {
-		return nil, e.getErr
-	}
-	if e.getResult != nil {
-		return e.getResult, nil
-	}
-	return e.ManagementClusterContentCRUD.Get(ctx, resourceID)
-}
-
-func (e *errorInjectingMCCCRUDForNodePool) Replace(ctx context.Context, obj *api.ManagementClusterContent, opts *azcosmos.ItemOptions) (*api.ManagementClusterContent, error) {
-	if e.replaceErr != nil {
-		return nil, e.replaceErr
-	}
-	return e.ManagementClusterContentCRUD.Replace(ctx, obj, opts)
-}
-
-var _ database.ManagementClusterContentCRUD = &errorInjectingMCCCRUDForNodePool{}
-
-func TestReadAndPersistNodePoolScopedMaestroReadonlyBundlesContentSyncer_calculateManagementClusterContentFromMaestroBundle(t *testing.T) {
-	nodepoolResourceID := api.Must(azcorearm.ParseResourceID("/subscriptions/sub/resourceGroups/rg/providers/Microsoft.RedHatOpenShift/hcpOpenShiftClusters/cluster/nodePools/test-nodepool"))
-	nodepool := &api.HCPOpenShiftClusterNodePool{
-		TrackedResource: arm.TrackedResource{
-			Resource: arm.Resource{
-				ID:   nodepoolResourceID,
-				Name: "test-nodepool",
-			},
-		},
-	}
-	ref := &api.MaestroBundleReference{
-		Name:                        api.MaestroBundleInternalNameReadonlyHypershiftNodePool,
-		MaestroAPIMaestroBundleName: "bundle-name",
-	}
-
-	np := hsv1beta1.NodePool{TypeMeta: metav1.TypeMeta{APIVersion: "hypershift.openshift.io/v1beta1", Kind: "NodePool"}, ObjectMeta: metav1.ObjectMeta{Name: "np1", Namespace: "ns1"}}
-	npJSONBytes, err := json.Marshal(np)
-	require.NoError(t, err)
-	validNPJSON := string(npJSONBytes)
-
-	tests := []struct {
-		name            string
-		maestroGet      func(*maestro.MockClient)
-		wantDegraded    bool
-		wantKubeContent bool
-		wantErr         bool
-		errSub          string
-	}{
-		{
-			name: "bundle not found - desired with degraded condition",
-			maestroGet: func(m *maestro.MockClient) {
-				m.EXPECT().Get(gomock.Any(), "bundle-name", gomock.Any()).Return(nil, k8serrors.NewNotFound(schema.GroupResource{}, "bundle-name"))
-			},
-			wantDegraded:    true,
-			wantKubeContent: false,
-		},
-		{
-			name: "maestro bundle get error - returns error",
-			maestroGet: func(m *maestro.MockClient) {
-				m.EXPECT().Get(gomock.Any(), "bundle-name", gomock.Any()).Return(nil, fmt.Errorf("connection error"))
-			},
-			wantErr: true,
-			errSub:  "failed to get Maestro Bundle",
-		},
-		{
-			name: "bundle has invalid status feedback - desired with degraded",
-			maestroGet: func(m *maestro.MockClient) {
-				b := &workv1.ManifestWork{
-					Status: workv1.ManifestWorkStatus{
-						ResourceStatus: workv1.ManifestResourceStatus{
-							Manifests: []workv1.ManifestCondition{
-								{ResourceMeta: workv1.ManifestResourceMeta{}, StatusFeedbacks: workv1.StatusFeedbackResult{Values: []workv1.FeedbackValue{}}, Conditions: []metav1.Condition{}},
-							},
-						},
-					},
-				}
-				m.EXPECT().Get(gomock.Any(), "bundle-name", gomock.Any()).Return(b, nil)
-			},
-			wantDegraded:    true,
-			wantKubeContent: false,
-		},
-		{
-			name: "success - desired with kube content",
-			maestroGet: func(m *maestro.MockClient) {
-				b := buildTestMaestroBundleWithStatusFeedback("bundle-name", "ns", validNPJSON)
-				m.EXPECT().Get(gomock.Any(), "bundle-name", gomock.Any()).Return(b, nil)
-			},
-			wantDegraded:    false,
-			wantKubeContent: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			ctrl := gomock.NewController(t)
-			mockMaestro := maestro.NewMockClient(ctrl)
-			tt.maestroGet(mockMaestro)
-
-			got, err := calculateManagementClusterContentFromMaestroBundle(context.Background(), nodepool.ID, ref, mockMaestro)
-			if tt.wantErr {
-				require.Error(t, err)
-				assert.Contains(t, err.Error(), tt.errSub)
-			} else {
-				require.NoError(t, err)
-				require.NotNil(t, got)
-				assert.Equal(t, tt.wantKubeContent, got.Status.KubeContent != nil && len(got.Status.KubeContent.Items) > 0)
-				hasDegradedTrue := controllerutils.IsConditionTrue(got.Status.Conditions, "Degraded")
-				assert.Equal(t, tt.wantDegraded, hasDegradedTrue)
-			}
-		})
-	}
-}
-
-func TestReadAndPersistNodePoolScopedMaestroReadonlyBundlesContentSyncer_readAndPersistMaestroBundleContent(t *testing.T) {
-	ctx := context.Background()
-	nodepoolResourceID := api.Must(azcorearm.ParseResourceID("/subscriptions/sub/resourceGroups/rg/providers/Microsoft.RedHatOpenShift/hcpOpenShiftClusters/cluster/nodePools/test-nodepool"))
-	nodepool := &api.HCPOpenShiftClusterNodePool{
-		TrackedResource: arm.TrackedResource{
-			Resource: arm.Resource{
-				ID:   nodepoolResourceID,
-				Name: "test-nodepool",
-			},
-		},
-	}
-	bundleInternalName := api.MaestroBundleInternalNameReadonlyHypershiftNodePool
-	ref := &api.MaestroBundleReference{
-		Name:                        bundleInternalName,
-		MaestroAPIMaestroBundleName: "bundle-name",
-	}
-
-	np := hsv1beta1.NodePool{TypeMeta: metav1.TypeMeta{APIVersion: "hypershift.openshift.io/v1beta1", Kind: "NodePool"}, ObjectMeta: metav1.ObjectMeta{Name: "np1", Namespace: "ns1"}}
-	npJSONBytes, err := json.Marshal(np)
-	require.NoError(t, err)
-	validNPJSON := string(npJSONBytes)
-
-	// Node-pool–scoped ManagementClusterContents are nested under the node pool ARM resource.
-	const mccContainerSub = "sub"
-	const mccContainerRG = "rg"
-	const mccContainerCluster = "cluster"
-	const mccNodePoolName = "test-nodepool"
-	existingMCCResourceIDStr := fmt.Sprintf("/subscriptions/%s/resourceGroups/%s/providers/Microsoft.RedHatOpenShift/hcpOpenShiftClusters/%s/nodePools/%s/managementClusterContents/%s",
-		mccContainerSub, mccContainerRG, mccContainerCluster, mccNodePoolName, bundleInternalName)
-
-	t.Run("creates new ManagementClusterContent when not found", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockMaestro := maestro.NewMockClient(ctrl)
-		b := buildTestMaestroBundleWithStatusFeedback("bundle-name", "ns", validNPJSON)
-		mockMaestro.EXPECT().Get(gomock.Any(), "bundle-name", gomock.Any()).Return(b, nil)
-
-		mockDB := databasetesting.NewMockDBClient()
-		mccCRUD := mockDB.HCPClusters(mccContainerSub, mccContainerRG).NodePools(mccContainerCluster).ManagementClusterContents(mccNodePoolName)
-
-		err := readAndPersistMaestroReadonlyBundleContent(ctx, nodepool.ID, ref, mockMaestro, mccCRUD)
-		require.NoError(t, err)
-
-		got, err := mccCRUD.Get(ctx, string(bundleInternalName))
-		require.NoError(t, err)
-		require.NotNil(t, got)
-		require.NotNil(t, got.Status.KubeContent)
-		require.Len(t, got.Status.KubeContent.Items, 1)
-	})
-
-	t.Run("replaces existing when content changed", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockMaestro := maestro.NewMockClient(ctrl)
-		b := buildTestMaestroBundleWithStatusFeedback("bundle-name", "ns", validNPJSON)
-		mockMaestro.EXPECT().Get(gomock.Any(), "bundle-name", gomock.Any()).Return(b, nil)
-
-		mockDB := databasetesting.NewMockDBClient()
-		mccCRUD := mockDB.HCPClusters(mccContainerSub, mccContainerRG).NodePools(mccContainerCluster).ManagementClusterContents(mccNodePoolName)
-		existingRID := api.Must(azcorearm.ParseResourceID(existingMCCResourceIDStr))
-		existing := &api.ManagementClusterContent{
-			CosmosMetadata: api.CosmosMetadata{ResourceID: existingRID},
-			ResourceID:     *existingRID,
-			Status:         api.ManagementClusterContentStatus{KubeContent: &metav1.List{Items: []runtime.RawExtension{}}},
-		}
-		_, err := mccCRUD.Create(ctx, existing, nil)
-		require.NoError(t, err)
-
-		err = readAndPersistMaestroReadonlyBundleContent(ctx, nodepool.ID, ref, mockMaestro, mccCRUD)
-		require.NoError(t, err)
-
-		got, err := mccCRUD.Get(ctx, string(bundleInternalName))
-		require.NoError(t, err)
-		require.NotNil(t, got.Status.KubeContent)
-		require.Len(t, got.Status.KubeContent.Items, 1)
-	})
-
-	t.Run("keeps existing kube content when desired has no content (degraded)", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockMaestro := maestro.NewMockClient(ctrl)
-		// Bundle with no valid status feedback -> desired has no KubeContent
-		b := &workv1.ManifestWork{
-			Status: workv1.ManifestWorkStatus{
-				ResourceStatus: workv1.ManifestResourceStatus{
-					Manifests: []workv1.ManifestCondition{
-						{ResourceMeta: workv1.ManifestResourceMeta{}, StatusFeedbacks: workv1.StatusFeedbackResult{Values: []workv1.FeedbackValue{}}, Conditions: []metav1.Condition{}},
-					},
-				},
-			},
-		}
-		mockMaestro.EXPECT().Get(gomock.Any(), "bundle-name", gomock.Any()).Return(b, nil)
-
-		mockDB := databasetesting.NewMockDBClient()
-		mccCRUD := mockDB.HCPClusters(mccContainerSub, mccContainerRG).NodePools(mccContainerCluster).ManagementClusterContents(mccNodePoolName)
-		existingRID := api.Must(azcorearm.ParseResourceID(existingMCCResourceIDStr))
-		existingContent := &metav1.List{Items: []runtime.RawExtension{{Raw: []byte(`{}`)}}}
-		existing := &api.ManagementClusterContent{
-			CosmosMetadata: api.CosmosMetadata{ResourceID: existingRID},
-			ResourceID:     *existingRID,
-			Status:         api.ManagementClusterContentStatus{KubeContent: existingContent},
-		}
-		_, err := mccCRUD.Create(ctx, existing, nil)
-		require.NoError(t, err)
-
-		err = readAndPersistMaestroReadonlyBundleContent(ctx, nodepool.ID, ref, mockMaestro, mccCRUD)
-		require.NoError(t, err)
-
-		got, err := mccCRUD.Get(ctx, string(bundleInternalName))
-		require.NoError(t, err)
-		require.NotNil(t, got.Status.KubeContent)
-		assert.Equal(t, existingContent.Items[0].Raw, got.Status.KubeContent.Items[0].Raw)
-	})
-
-	t.Run("no replace occurs when content has not changed", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockMaestro := maestro.NewMockClient(ctrl)
-		b := buildTestMaestroBundleWithStatusFeedback("bundle-name", "ns", validNPJSON)
-		mockMaestro.EXPECT().Get(gomock.Any(), "bundle-name", gomock.Any()).Return(b, nil).Times(2)
-
-		mockDB := databasetesting.NewMockDBClient()
-		mccCRUD := mockDB.HCPClusters(mccContainerSub, mccContainerRG).NodePools(mccContainerCluster).ManagementClusterContents(mccNodePoolName)
-		existingRID := api.Must(azcorearm.ParseResourceID(existingMCCResourceIDStr))
-		desired, err := calculateManagementClusterContentFromMaestroBundle(ctx, nodepool.ID, ref, mockMaestro)
-		require.NoError(t, err)
-		existing := &api.ManagementClusterContent{
-			CosmosMetadata: api.CosmosMetadata{ResourceID: existingRID},
-			ResourceID:     *existingRID,
-			Status:         desired.Status,
-		}
-		_, err = mccCRUD.Create(ctx, existing, nil)
-		require.NoError(t, err)
-
-		err = readAndPersistMaestroReadonlyBundleContent(ctx, nodepool.ID, ref, mockMaestro, mccCRUD)
-		require.NoError(t, err)
-
-		got, err := mccCRUD.Get(ctx, string(bundleInternalName))
-		require.NoError(t, err)
-		require.NotNil(t, got.Status.KubeContent)
-		require.Len(t, got.Status.KubeContent.Items, 1)
-	})
-
-	t.Run("error occurs when object has been modified in Cosmos since we retrieved it", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockMaestro := maestro.NewMockClient(ctrl)
-		b := buildTestMaestroBundleWithStatusFeedback("bundle-name", "ns", validNPJSON)
-		mockMaestro.EXPECT().Get(gomock.Any(), "bundle-name", gomock.Any()).Return(b, nil)
-
-		existingRID := api.Must(azcorearm.ParseResourceID(existingMCCResourceIDStr))
-		existingDoc := &api.ManagementClusterContent{
-			CosmosMetadata: api.CosmosMetadata{ResourceID: existingRID},
-			ResourceID:     *existingRID,
-			Status:         api.ManagementClusterContentStatus{KubeContent: &metav1.List{Items: []runtime.RawExtension{{Raw: []byte(validNPJSON)}}}},
-		}
-
-		mockDB := databasetesting.NewMockDBClient()
-		baseMccCRUD := mockDB.HCPClusters(mccContainerSub, mccContainerRG).NodePools(mccContainerCluster).ManagementClusterContents(mccNodePoolName)
-		mccCRUD := &errorInjectingMCCCRUDForNodePool{
-			ManagementClusterContentCRUD: baseMccCRUD,
-			getResult:                    existingDoc,
-			replaceErr:                   databasetesting.NewPreconditionFailedError(),
-		}
-
-		err := readAndPersistMaestroReadonlyBundleContent(ctx, nodepool.ID, ref, mockMaestro, mccCRUD)
-		require.Error(t, err)
-		assert.Contains(t, err.Error(), "failed to replace ManagementClusterContent")
-		assert.True(t, database.IsResponseError(err, http.StatusPreconditionFailed), "expected 412 Precondition Failed")
-	})
-
-	t.Run("error occurs when managementClusterContentsDBClient.Get fails", func(t *testing.T) {
-		ctrl := gomock.NewController(t)
-		mockMaestro := maestro.NewMockClient(ctrl)
-		b := buildTestMaestroBundleWithStatusFeedback("bundle-name", "ns", validNPJSON)
-		mockMaestro.EXPECT().Get(gomock.Any(), "bundle-name", gomock.Any()).Return(b, nil)
-
-		mockDB := databasetesting.NewMockDBClient()
-		baseMccCRUD := mockDB.HCPClusters(mccContainerSub, mccContainerRG).NodePools(mccContainerCluster).ManagementClusterContents(mccNodePoolName)
-		mccCRUD := &errorInjectingMCCCRUDForNodePool{
-			ManagementClusterContentCRUD: baseMccCRUD,
-			getErr:                       fmt.Errorf("cosmos connection error"),
-		}
-
-		err := readAndPersistMaestroReadonlyBundleContent(ctx, nodepool.ID, ref, mockMaestro, mccCRUD)
-		require.Error(t, err)
-		assert.Contains(t, err.Error(), "failed to get ManagementClusterContent")
-		assert.Contains(t, err.Error(), "cosmos connection error")
-	})
 }
 
 func TestReadAndPersistNodePoolScopedMaestroReadonlyBundlesContentSyncer_SyncOnce_NodePoolNotFound(t *testing.T) {
